+++ src/Stormancer.Plugins/Profile/Stormancer.Server.Plugins.Profile/CustomPartBuilder.cs
@@ -1,4 +1,4 @@
-// MIT License
+﻿// MIT License
 //
 // Copyright (c) 2019 Stormancer
 //
@@ -20,42 +20,30 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using Stormancer.Core;
-using Stormancer.Plugins;
-using Stormancer.Server.Plugins.GameFinder;
+using Stormancer.Diagnostics;
 using System;
-using System.IO;
-using System.Reactive.Linq;
-using System.Threading;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
 using System.Threading.Tasks;
 
-namespace Stormancer.Server.Plugins.GameFinder
+namespace Stormancer.Server.Plugins.Profile
 {
-    class GameFinderProxy : IGameFinder
+    class CustomPartBuilder : IProfilePartBuilder
     {
-        public readonly ISceneHost _scene;
-        public readonly ISerializer _serializer;
+        private readonly IEnumerable<ICustomProfilePart> parts;
+        private readonly ILogger logger;
 
-        public GameFinderProxy(ISceneHost scene,
-            ISerializer serializer)
+        public CustomPartBuilder(IEnumerable<ICustomProfilePart> parts,
+            ILogger logger)
         {
-            _scene = scene;
-            _serializer = serializer;
+            this.parts = parts;
+            this.logger = logger;
         }
 
-        public async Task FindMatch(string sceneName, Models.Party gameFinderRequest, CancellationToken cancelationToken)
+        public Task GetProfiles(ProfileCtx ctx)
         {
-            var rpc = _scene.DependencyResolver.Resolve<RpcService>();
-            Action<Stream> writer = s =>
-            {
-                _serializer.Serialize(gameFinderRequest, s);
-            };
-            var packet = await rpc.Rpc(GameFinderController.FindGameS2SRoute, new MatchSceneFilter(sceneName), writer, PacketPriority.MEDIUM_PRIORITY, cancelationToken).Select(p =>
-            {
-                using (p) { }
-                return System.Reactive.Unit.Default;
-            }).LastOrDefaultAsync();
-
+            return parts.RunEventHandler(p => p.GetAsync(ctx), ex => logger.Log(LogLevel.Error, "profiles.customParts", "An error occured while processing custom part builders.", ex));
         }
     }
 }