+++ src/Stormancer.Plugins/Utilities/Stormancer.Server.Plugins.Utilities/Changelog.rst
@@ -9,6 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Removed cleanup cycle trace logs.
+
 2.0.3
 -----
 Changed